+++ index.html
@@ -10,8 +10,7 @@
 </head>
 
 <body>
-  <script src="js/function.js"></script>
-  <script src="js/main.js"></script>
+
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
@@ -228,6 +227,6 @@
       </div>
     </section>
   </template>
-
+  <script src="./js/main.js" type="module"></script>
 </body>
 </html>
